pyproject.toml
[tool.poetry]
name = "animals-polyclinic-main"
version = "0.1.0"
description = ""
authors = ["leoverem14"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.1"
aiogram = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
